Split eslint-config into per-plugin modules
Extract each plugin config into its own file under src/plugins/
with a uniform PluginFactory interface. The barrel (plugins/index.ts)
defines plugin ordering. configure() resolves option defaults into
ResolvedOptions and delegates to the plugin array.
